+++ NEWS.md
@@ -1,5 +1,11 @@
 # ncdfCF (development version)
 
+#### API
+- When selecting auxiliary coordinates on an axis, printing a `CFVariable` using that axis will print the name of the active coordinates for that axis and its units, if set.
+
+#### Code
+- Fixed NC_CHAR label/axis dimensions.
+
 # ncdfCF 0.8.1
 
 Patch release with bug fixes and some new functionality.